[BUGFIX] Do not unlink temporary files
When fetching files for local processing the FAL API defines
two ways to do so.

1. Fetch the file for modification
2. Fetch the file only for reading

In the first case the consumer of the API is responsible
for cleaning the temporary file after usage.

In the second case the driver is responsible
for cleaning up the temporary file at any point in time
when the driver considers it appropriate.

Since the tika extension always fetches the file only for reading,
and the API clearly states, that for such case NO modifications
must be done to the file (and removing it is clearly a pretty severe modification),
the local file must not be removed at all.

This is important for remote storage drivers that do some optimizations
to keep the local file cached during one request, to speed up operations on those
on subsequent calls for processing.