[TASK] Introduce TYPO3 HTTP client
The implementation to fetch website content based on the stream context and
method 'file_get_content()' is removed. Instead the TYPO3 build in
HTTP client is in use to access the Tika server. This client supports
PSR-18 and allows to use proxy information configured in TYPO3.

Internally the string representation replaced by the URI interface.
This allows an easier and better way to handle server URIs.

Fixes: #154