[TASK] Remove multiple TIKA_VERSION defs and use from composer.json only
The purpose of this change is the simplification of dev workflows.
By upgrading to newer versions of servers the devs were forced 
to change the version constrain on multiple places on this repo 
and other places like dev environments and so on.

Now the version is defined only on single place:
`composer.json` -> `extra.TYPO3-Solr.ext-tika.require.Tika`

By needs the required version can be fetched 
by following command examples:

* inside EXT:solr
** `composer tika:req:tika`
** `composer tika:req:solr`
* from root of TYPO3 project:
** `composer --working-dir=public/typo3conf/ext/tika tika:req:tika`
** `composer --working-dir=public/typo3conf/ext/tika tika:req:solr`

Relates: #185