Release 12.0.4
This is maintenance release for TYPO3 12 allowing PHP 8.4.

Note:
If you use the mata-data extractor with Solr 9.10.0+, it does not extract the data properly.
Please use the Tika Server or App instead. Or stay by Apache Solr 9.9.0.

Please read the release notes for more information:

* https://github.com/TYPO3-Solr/ext-tika/releases/tag/12.0.4
* https://docs.typo3.org/p/apache-solr-for-typo3/tika/12.0/en-us/Releases/12_0.html
* **Full Changelog**: https://github.com/TYPO3-Solr/ext-tika/compare/12.0.3...12.0.4

---

# How to Get Involved

There are many ways to get involved with Apache Solr for TYPO3:

* Submit bug reports and feature requests on [GitHub](https://github.com/TYPO3-Solr/ext-solr)
* Ask or help or answer questions in our [Slack channel](https://typo3.slack.com/messages/ext-solr/)
* Provide patches through Pull Request or review and comment on existing [Pull Requests](https://github.com/TYPO3-Solr/ext-solr/pulls)
* Go to [www.typo3-solr.com](http://www.typo3-solr.com) or call [dkd](http://www.dkd.de) to sponsor the ongoing development of Apache Solr for TYPO3

Support us by becoming an EB partner:

http://www.typo3-solr.com/en/contact/

or call:

+49 (0)69 - 2475218 0